requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
google-generativeai>=0.7.0
google>=0.1.0
PyJWT>=2.1.0
cryptography>=35.0.0
sqlalchemy-utils>=0.38.3
httpx>=0.23.0